Redraw dashboard less often to get up to maximum speed.
The previous implementation of the dashboard had a way to continously parse the
message buffer until no more were available, then draw the dashboard. This had
the nice effect of letting us definitely draw every time a message was received,
but if there was a high traffic load, buffer at least a few messages.

I'm not sure how we could do that with the current API, but it seems like it
would be useful.