Restrict openxc-gps to tracefiles.
We need to wait until the trace is compelted so we know to close the GPX file's
structure, so we wait for the trace playback to finish when creating a GPS log.

This isn't going to work well if you're using a VI directly, but I can't think
of a case when you would do that vs. recording a full OpenXC log.