Progress on Python CAN Translator
1) re-organized the generated data structures to be a bit more
pythonesque, wrapping signals inside message and state inside signal.

2) decodeCanMessage() and getMessageList() added to footer in generation
of signals.py

3) Grabbed canread.cpp from vi-firmware master and refactored into
python.  This is required to support any sort of obfustication if we
hope to release signals.pyc for specific vehicles.  Essentially,
canread.pyc and signals.pyc combine to compartmentalize the translation
of a raw can message to resulting OpenXC json completely.  A 3rd party
hardware designer would simply:

-import canread.pyc and signals.pyc for a specific vehicle
-use two entry points:  decodeCanMessage() and getMessageList().  The
second is to expose a list of CAN messages to setup any hardware
filters.

I wonder if this same approach could be used on the C side so Ford could
start releasing just a few obj files per vehicle for 3rd party hardware
designers instead of having to build a complete firmware for each
vehicle?

Status:
Currently, we have a prototype hardware device with a python interpreter
and an STN1100.  It is successfully setting up  CAN filtering (using a
call to getMessageList() to get appropriate message ids), then
processing each CAN message with a call into decodeCanMessage().  I'm
currently faking some CAN data that matches the sample json but I'd love
to get my hands on a json file for one of the cars we have access to or
even just a few CAN messages to do further testing and research.

Issues:

1) None of the write handler capability is in place really yet

2) for custom signal handlers, it is expected the producer of these
handlers will return json instead of a value (different that the C
version).  This is more flexible and also probably makes sense with
loose typing.

3) Not sure what exactly to do with message-level handlers (ie
HandleSteeringWheelMessage)  I don't see in the example where they are
supported in the current firmware.

4) Assumes Big Endian.